move PlayCard component into subfolder
